+++ source/plugins/correlation/ui/qml/Graphia/Plugins/Correlation/correlationplot.h
@@ -191,7 +191,7 @@ class CorrelationPlot : public QQuickPaintedItem, protected QCustomPlotColorProv
     Q_PROPERTY(QString colorGroupByAnnotationName MEMBER _colorGroupByAnnotationName
         WRITE setColorGroupByAnnotationName NOTIFY plotOptionsChanged)
 
-    Q_PROPERTY(std::vector<int> selectedColumns READ selectedColumns NOTIFY selectedColumnsChanged)
+    Q_PROPERTY(std::vector<size_t> selectedColumns READ selectedColumns NOTIFY selectedColumnsChanged)
 
     Q_PROPERTY(int elideLabelWidth MEMBER _elideLabelWidth WRITE setElideLabelWidth)
     Q_PROPERTY(bool showColumnNames MEMBER _showColumnNames WRITE setShowColumnNames NOTIFY plotOptionsChanged)
@@ -419,7 +419,7 @@ private:
     int plotMode() const;
     void setPlotMode(int plotModeInt);
 
-    std::vector<int> selectedColumns() const;
+    std::vector<size_t> selectedColumns() const;
 
     QStringList visibleColumnAnnotationNames() const;
     void setVisibleColumnAnnotationNames(const QStringList& columnAnnotations);
